Moved Add or remove enum to the controller package
