+++ src/main/java/com/qa/ims/controller/AddOrRemove.java
@@ -1,4 +1,4 @@
-package com.qa.ims.persistence.domain;
+package com.qa.ims.controller;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
