+++ src/main/java/com/qa/ims/controller/OrderController.java
@@ -9,7 +9,6 @@ import org.apache.logging.log4j.Logger;
 
 import com.qa.ims.persistence.dao.OrderDAO;
 import com.qa.ims.persistence.dao.OrderItemDAO;
-import com.qa.ims.persistence.domain.AddOrRemove;
 import com.qa.ims.persistence.domain.Order;
 import com.qa.ims.persistence.domain.OrderItem;
 import com.qa.ims.utils.Utils;
